# raffleease-openapi.yaml
openapi: 3.1.0
info:
  title: Raffleease API
  description: |
    Raffleease is a RESTful API designed to support the management of online raffles for animal rescue associations. The platform follows a **multi-tenant architecture**, where most endpoints are scoped under a specific association ID to ensure data isolation across organizations.
    
    The API implements **JWT-based authentication** with secure refresh token handling via **HTTP-only, same-site, secure cookies**. **Role-based access control** governs permissions across different user types within each association.
    
    Search endpoints support **advanced filtering, sorting, and pagination** mechanisms to allow flexible querying of resources. All API responses follow a **consistent, standardized format**, and **error handling** is uniform across the platform to provide clarity for integrators.
    
    This documentation should clearly outline available endpoints, authentication flows, security requirements, and expected request and response formats to help developers integrate with the Raffleease platform effectively.    

    ## Key Features
    
    ### Multi-Tenant Architecture
    Most endpoints are scoped under `/associations/{associationId}` to ensure data isolation and support multiple organizations.
    
    ### JWT-Based Authentication
    The API uses JWT access tokens for authentication and secure HTTP-only cookies to manage refresh tokens.
    
    ### Role-Based Access Control
...
    
...
    
...
    
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/associations/{associationId}/raffles:
    post:
      tags:
        - Raffles
      summary: Create new raffle
      description: |
        Creates a new raffle within an association with images and ticket configuration.
        
        **Authorization:**
        - Only ADMIN and MEMBER users can create raffles
        - User must belong to the target association
        
        **Business Rules:**
        - Requires at least 1 image, maximum 10 images
        - Start date and end date must be in the future
        - End date must be after start date
        - Ticket amount must be positive
        - Ticket price must be positive
        - Lower limit must be non-negative
        
        **Image Association:**
        - Uses pending images uploaded by the authenticated user
        - Images are moved from "PENDING" to "ACTIVE" status
        - Images are associated with the newly created raffle
        
        **Initial Status:** Raffle is created in "PENDING" status
      operationId: createRaffle
      parameters:
        - $ref: '#/components/parameters/associationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RaffleCreateRequest'
      responses:
        '201':
          description: Raffle created successfully
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
            Location:
              description: URL of the created raffle
              schema:
                type: string
                format: uri
              example: "https://api.raffleease.es/v1/associations/456/raffles/789"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RaffleResponse'
              example:
                success: true
                message: "New raffle created successfully"
                timestamp: "2024-01-15T10:30:00Z"
                data:
                  id: 789
                  associationId: 456
                  title: "Help Rescue Dogs Find Homes"
                  status: "PENDING"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags:
        - Raffles
      summary: Search raffles
      description: |
        Searches for raffles within an association with filtering, sorting, and pagination.
        
        **Public Endpoint:** No authentication required
        
        **Filtering:**
        - Filter by title (text search)
        - Filter by status (PENDING, ACTIVE, PAUSED, COMPLETED)
        
        **Sorting:**
        - Sort by any raffle field (createdAt, endDate, title, etc.)
        - Supports ascending and descending order
        
        **Pagination:**
        - Page-based pagination with configurable page size
        - Returns total count and page metadata
      operationId: searchRaffles
      security: []  # Public endpoint
      parameters:
        - $ref: '#/components/parameters/associationId'
        - name: title
          in: query
          required: false
          description: Filter by raffle title (case-insensitive partial match)
          schema:
            type: string
          example: "rescue dogs"
        - name: status
          in: query
          required: false
          description: Filter by raffle status
          schema:
            $ref: '#/components/schemas/RaffleStatus'
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/sizeParam'
        - $ref: '#/components/parameters/sortParam'
      responses:
        '200':
          description: Raffles retrieved successfully
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PageResponse'
                          - type: object
                            properties:
                              content:
                                type: array
                                items:
                                  $ref: '#/components/schemas/RaffleResponse'
              example:
                success: true
                message: "Raffles retrieved successfully"
                timestamp: "2024-01-15T10:30:00Z"
                data:
                  content:
                    - id: 789
                      title: "Help Rescue Dogs Find Homes"
                      status: "ACTIVE"
                      endDate: "2024-02-15T18:00:00Z"
                  totalElements: 25
                  totalPages: 3
                  size: 20
                  number: 0
                  first: true
                  last: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/associations/{associationId}/raffles/{raffleId}:
    get:
      tags:
        - Raffles
      summary: Get raffle details
      description: |
        Retrieves detailed information about a specific raffle including statistics.
        
        **Authorization:**
        - Requires authentication
        - User must have access to the association
        
        **Response includes:**
        - Complete raffle information
        - Real-time statistics (sold tickets, revenue, participants, etc.)
        - Image metadata and URLs
        - Status and completion information
      operationId: getRaffle
      parameters:
        - $ref: '#/components/parameters/associationId'
        - name: raffleId
          in: path
          required: true
          description: The unique identifier of the raffle
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 789
      responses:
        '200':
          description: Raffle retrieved successfully
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RaffleResponse'
              example:
                success: true
                message: "Raffle retrieved successfully"
                timestamp: "2024-01-15T10:30:00Z"
                data:
                  id: 789
                  associationId: 456
                  title: "Help Rescue Dogs Find Homes"
                  description: "Help us raise funds to provide shelter and care for rescued dogs."
                  status: "ACTIVE"
                  ticketPrice: 5.00
                  totalTickets: 1000
                  statistics:
                    soldTickets: 250
                    revenue: 1250.00
                    participants: 180
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Raffles
      summary: Edit raffle
      description: |
        Updates an existing raffle with new information.
        
        **Authorization:**
        - Only ADMIN and MEMBER users can edit raffles
        - User must belong to the association
        
        **Business Rules:**
        - Can only edit certain fields (title, description, endDate, images, ticketPrice, totalTickets)
        - End date must be in the future if provided
        - Images must be between 1-10 if provided
        - Ticket price must be positive if provided
        - Total tickets must be positive if provided
        
        **Partial Updates:**
        - All fields are optional
        - Only provided fields are updated
        - Images are replaced entirely if provided
        
        **Status Considerations:**
        - Some fields may have restrictions based on current raffle status
        - Cannot reduce total tickets below already sold tickets
      operationId: editRaffle
      parameters:
        - $ref: '#/components/parameters/associationId'
        - name: raffleId
          in: path
          required: true
          description: The unique identifier of the raffle to edit
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 789
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RaffleEditRequest'
      responses:
        '200':
          description: Raffle edited successfully
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RaffleResponse'
              example:
                success: true
                message: "Raffle edited successfully"
                timestamp: "2024-01-15T10:30:00Z"
                data:
                  id: 789
                  title: "Updated: Help Rescue Dogs Find Homes"
                  status: "ACTIVE"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Raffles
      summary: Delete raffle
      description: |
        Deletes a raffle from the system.
        
        **Authorization:**
        - Only ADMIN and MEMBER users can delete raffles
        - User must belong to the association
        
        **Business Rules:**
        - Can only delete raffles in "PENDING" status
        - Raffles with any other status cannot be deleted
        - This is a hard delete operation
        
        **Side Effects:**
        - Associated images are marked for deletion
        - Associated tickets are removed
        - Operation is irreversible
      operationId: deleteRaffle
      parameters:
        - $ref: '#/components/parameters/associationId'
        - name: raffleId
          in: path
          required: true
          description: The unique identifier of the raffle to delete
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 789
      responses:
        '204':
          description: Raffle deleted successfully
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
        '400':
          description: Bad Request - Raffle cannot be deleted (not in PENDING status)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "Only raffles in 'PENDING' state can be deleted."
                timestamp: "2024-01-15T10:30:00Z"
                statusCode: 400
                statusText: "Bad Request"
                code: "BUSINESS_ERROR"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/associations/{associationId}/raffles/{raffleId}/status:
    patch:
      tags:
        - Raffles
      summary: Update raffle status
      description: |
        Updates the status of a raffle with complex business rule validation.
        
        **Authorization:**
        - Only ADMIN and MEMBER users can update raffle status
        - User must belong to the association
        
        **Status Transitions:**
        
        **PENDING → ACTIVE:**
        - Validates end date is at least 24 hours in the future
        - Sets start date to current time
        
        **ACTIVE → PAUSED:**
        - Direct transition allowed
        - Temporarily stops ticket sales
        
        **PAUSED → ACTIVE:**
        - Validates end date is at least 24 hours in the future
        - Resumes ticket sales
        
        **ACTIVE/PAUSED → COMPLETED:**
        - Manual completion
        - Sets completion reason to "MANUALLY_COMPLETED"
        - Sets completed timestamp
        
        **COMPLETED → ACTIVE (Reactivation):**
        - Cannot reactivate if raffle has a winner
        - End date must be at least 24 hours in the future
        - Must have available tickets (total > sold)
        - Clears completion reason and timestamp
        
        **Invalid Transitions:**
        - Cannot revert to PENDING status
        - Other invalid transitions return business errors
        
        **Automatic Completion:**
        - System automatically completes raffles when all tickets are sold
        - System automatically completes raffles when end date is reached
      operationId: updateRaffleStatus
      parameters:
        - $ref: '#/components/parameters/associationId'
        - name: raffleId
          in: path
          required: true
          description: The unique identifier of the raffle
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 789
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusUpdateRequest'
      responses:
        '200':
          description: Raffle status updated successfully
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RaffleResponse'
              example:
                success: true
                message: "Raffle status updated successfully"
                timestamp: "2024-01-15T10:30:00Z"
                data:
                  id: 789
                  status: "ACTIVE"
                  startDate: "2024-01-15T10:30:00Z"
        '400':
          description: Bad Request - Invalid status transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidTransition:
                  value:
                    success: false
                    message: "Cannot revert to 'PENDING' state."
                    timestamp: "2024-01-15T10:30:00Z"
                    statusCode: 400
                    statusText: "Bad Request"
                    code: "BUSINESS_ERROR"
                endDateTooSoon:
                  value:
                    success: false
                    message: "The end date of the raffle must be at least one day after the current date to reactivate"
                    timestamp: "2024-01-15T10:30:00Z"
                    statusCode: 400
                    statusText: "Bad Request"
                    code: "BUSINESS_ERROR"
                hasWinner:
                  value:
                    success: false
                    message: "Cannot reactivate a raffle that already has a winner"
                    timestamp: "2024-01-15T10:30:00Z"
                    statusCode: 400
                    statusText: "Bad Request"
                    code: "BUSINESS_ERROR"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for API authentication
  parameters:
    associationId:
      name: associationId
      in: path
      required: true
      description: The unique identifier of the association for multi-tenant resource access
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 123
...
  headers:
    X-Rate-Limit-Remaining:
      description: Number of requests remaining in current rate limit window
      schema:
        type: integer
    X-Rate-Limit-Reset:
      description: Time when rate limit window resets (Unix timestamp)
      schema:
        type: integer
    X-Rate-Limit-Limit:
      description: Total number of requests allowed in rate limit window
      schema:
        type: integer
  schemas:
...
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              enum: [true]
            data:
              description: The response payload containing the requested data
              oneOf:
                - type: object
                - type: array
                - type: string
                - type: number
                - type: boolean
                - type: "null"
      example:
        success: true
        message: "Data retrieved successfully"
        timestamp: "2024-01-15T10:30:00Z"
        data:
          id: 123
          name: "Example"
...
    RaffleCreateRequest:
      type: object
      required:
        - title
        - description
        - endDate
        - images
        - ticketsInfo
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: The title of the raffle
          example: "Help Rescue Dogs Find Homes"
        description:
          type: string
          minLength: 1
          maxLength: 5000
          description: The description of the raffle
          example: "Help us raise funds to provide shelter and care for rescued dogs in our community. Every ticket sold helps provide food, medical care, and shelter for animals in need."
        startDate:
          type: string
          format: date-time
          description: |
            The start date of the raffle (must be in the future).
            If not provided, will be set when raffle status changes to ACTIVE.
          example: "2024-01-20T09:00:00Z"
        endDate:
          type: string
          format: date-time
          description: The end date of the raffle (must be in the future and after start date)
          example: "2024-02-15T18:00:00Z"
        images:
          type: array
          items:
            $ref: '#/components/schemas/ImageDTO'
          minItems: 1
          maxItems: 10
          description: Array of images for the raffle (uses pending images uploaded by the user)
        ticketsInfo:
          $ref: '#/components/schemas/TicketsCreateInfo'
...
    RaffleResponse:
      type: object
      required:
        - id
        - associationId
        - title
        - description
        - status
        - ticketPrice
        - totalTickets
        - firstTicketNumber
        - images
        - statistics
        - createdAt
        - updatedAt
        - endDate
      properties:
        id:
          type: integer
          format: int64
          description: The unique identifier of the raffle
          example: 789
        associationId:
          type: integer
          format: int64
          description: The unique identifier of the association
          example: 456
        title:
          type: string
          description: The title of the raffle
          example: "Help Rescue Dogs Find Homes"
        description:
          type: string
          description: The description of the raffle
          example: "Help us raise funds to provide shelter and care for rescued dogs."
        status:
          $ref: '#/components/schemas/RaffleStatus'
        ticketPrice:
          type: number
          format: double
          description: The price of each ticket
          example: 5.00
        totalTickets:
          type: integer
          format: int64
          description: The total number of tickets available for the raffle
          example: 1000
        firstTicketNumber:
          type: integer
          format: int64
          description: The first ticket number in the raffle sequence
          example: 1
        images:
          type: array
          items:
            $ref: '#/components/schemas/ImageDTO'
          description: Array of images associated with the raffle
        completionReason:
          $ref: '#/components/schemas/CompletionReason'
        winningTicketId:
          type: integer
          format: int64
          description: The ID of the winning ticket (if raffle is completed and has a winner)
          example: 12345
          nullable: true
        statistics:
          $ref: '#/components/schemas/RaffleStatistics'
        createdAt:
          type: string
          format: date-time
          description: When the raffle was created
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: When the raffle was last updated
          example: "2024-01-15T11:45:00Z"
        startDate:
          type: string
          format: date-time
          description: When the raffle started (set when status changes to ACTIVE)
          example: "2024-01-20T09:00:00Z"
          nullable: true
        endDate:
          type: string
          format: date-time
          description: When the raffle ends
          example: "2024-02-15T18:00:00Z"
        completedAt:
          type: string
          format: date-time
          description: When the raffle was completed (if applicable)
          example: "2024-02-15T18:00:00Z"
          nullable: true
...
  responses:
...
    BadRequest:
      description: Bad Request - The request was invalid or malformed
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Invalid request parameters"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 400
            statusText: "Bad Request"
            code: "BAD_REQUEST"
    Unauthorized:
      description: Unauthorized - Authentication required or token invalid
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Authentication required"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 401
            statusText: "Unauthorized"
            code: "UNAUTHORIZED"
    Forbidden:
      description: Forbidden - Insufficient permissions to access the resource
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Access denied - insufficient permissions"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 403
            statusText: "Forbidden"
            code: "ACCESS_DENIED"
...
    ValidationError:
      description: Unprocessable Entity - Validation failed
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            success: false
            message: "Validation failed"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 400
            statusText: "Bad Request"
            code: "VALIDATION_ERROR"
            errors:
              email: "INVALID_FORMAT"
              phoneNumber: "REQUIRED"
    TooManyRequests:
      description: Too Many Requests - Rate limit exceeded
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
        Retry-After:
          description: Number of seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Rate limit exceeded. Please try again later."
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 429
            statusText: "Too Many Requests"
            code: "RATE_LIMIT_EXCEEDED"
    InternalServerError:
      description: Internal Server Error - An unexpected error occurred
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "An unexpected error occurred"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 500
            statusText: "Internal Server Error"
            code: "UNEXPECTED_ERROR" 